+++ automation/wiki_frontend/assistant_ui_app/src/App.tsx
@@ -12,7 +12,6 @@ import { SpotliteBoard } from "./domains/mission/components/SpotliteBoard";
 import { PaperclipStudio } from "./domains/paperclip/components/PaperclipStudio";
 import { WikiWorkspace } from "./domains/wiki/components/WikiWorkspace";
 import { SurfaceMigrationFallback } from "./surfaces/SurfaceMigrationFallback";
-import { ToastCenterProvider } from "./components/surface/ToastCenter";
 import {
   getPrimarySurfaceDefinition,
   getSurfaceDefinition,
@@ -237,15 +236,13 @@ export function App() {
   })();
 
   return (
-    <ToastCenterProvider>
-      <ProductFrame
-        activeSurface={surface}
-        onPrimaryChange={selectPrimarySurface}
-        onSurfaceChange={selectSurface}
-        workspaceLabel={chatContext.workspace}
-      >
-        {content}
-      </ProductFrame>
-    </ToastCenterProvider>
+    <ProductFrame
+      activeSurface={surface}
+      onPrimaryChange={selectPrimarySurface}
+      onSurfaceChange={selectSurface}
+      workspaceLabel={chatContext.workspace}
+    >
+      {content}
+    </ProductFrame>
   );
 }